minor improvements
