+++ src/Components/BackOffice/OrganizationData/OrganizationData.jsx
@@ -1,0 +1,8 @@
+
+export function OrganizationData() {
+    return (
+        <section className="organization-info">
+            
+        </section>
+    )
+}
