+++ src/hooks/useDrawer.js
@@ -1,0 +1,17 @@
+import { useState } from "react";
+
+export function useDrawer() {
+    const [open, setOpen] = useState(false);
+    const showDrawer = () => {
+		setOpen(true)
+	}
+	const onClose = () => {
+		setOpen(false)
+	}
+
+    return {
+        open,
+        showDrawer,
+        onClose
+    }
+}
