+++ src/pages/backoffice/BackOfficeDashboard.jsx
@@ -1,51 +1,14 @@
 import { faComments, faImage, faNewspaper, faRectangleList } from "@fortawesome/free-regular-svg-icons";
 import { faListCheck, faSitemap, faUserPen, faUsers } from "@fortawesome/free-solid-svg-icons";
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
-import { BackOfficeLayout, SectionCard } from '../Components/BackOffice'
-import { MenuOutlined  } from '@ant-design/icons'
-import { Drawer } from 'antd';
-import { useState } from "react";
-import { Link } from "react-router-dom";
+import { BackOfficeLayout, SectionCard } from '../../Components/BackOffice'
+import BackOfficeNavbar from "../../Components/BackOffice/BackOfficeNavbar";
 
 export function BackOfficeDashboard() {
-    const [open, setOpen] = useState(false);
-    const [placement, setPlacement] = useState('right');
-    const showDrawer = () => {
-		setOpen(true)
-	}
-	const onClose = () => {
-		setOpen(false)
-	}
+
 	return (
 		<BackOfficeLayout>
-			<nav className="backoffice-navbar">
-				<MenuOutlined
-					style={{
-						fontSize: '2rem',
-						cursor: 'pointer',
-					}}
-                    onClick={showDrawer}
-				/>
-			</nav>
-			<Drawer
-				title="Asi somos ❤"
-				placement={placement}
-				width={500}
-				onClose={onClose}
-				open={open}
-                style={{
-                    fontSize: '1.5rem',
-                }}
-			>
-                <div className="drawer-links">
-                    <Link to={`/backoffice/usuarios`}>
-                        Cerrar sesion
-                    </Link>
-                    <Link to={`/backoffice/usuarios`}>
-                        Ir al sitio
-                    </Link>
-                </div>
-			</Drawer>
+			<BackOfficeNavbar />
 			<SectionCard
 				title={'Productos'}
 				image={() => <FontAwesomeIcon icon={faNewspaper} size="6x" />}
