+++ src/pages/backoffice/OrganizationInfo.jsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+export default function OrganizationInfo() {
+    return (
+        <main className=''>
+            
+        </main>
+    )
+}
