+++ src/sass/main.scss
@@ -2,7 +2,7 @@
 @import './variables';
 @import './components/sectionCard';
 @import './components/backOfficeLayout';
-
+@import './components/organizationData';
 
 body{
     min-height: 100vh;
